refactor(admin): organize admin pages into feature folders
