+++ src/pages/admin/categories/CategoryAddEdit.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
-import { useCategoryRepository } from '../../hooks/useCategoryRepository';
-import { useChartOfAccountRepository } from '../../hooks/useChartOfAccountRepository';
-import { Category, CategoryType } from '../../models/category.model';
-import { Card, CardHeader, CardContent, CardFooter } from '../../components/ui2/card';
-import { Input } from '../../components/ui2/input';
-import { Button } from '../../components/ui2/button';
-import { Textarea } from '../../components/ui2/textarea';
-import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from '../../components/ui2/select';
-import { Switch } from '../../components/ui2/switch';
-import BackButton from '../../components/BackButton';
+import { useCategoryRepository } from '../../../hooks/useCategoryRepository';
+import { useChartOfAccountRepository } from '../../../hooks/useChartOfAccountRepository';
+import { Category, CategoryType } from '../../../models/category.model';
+import { Card, CardHeader, CardContent, CardFooter } from '../../../components/ui2/card';
+import { Input } from '../../../components/ui2/input';
+import { Button } from '../../../components/ui2/button';
+import { Textarea } from '../../../components/ui2/textarea';
+import { Select, SelectTrigger, SelectValue, SelectContent, SelectItem } from '../../../components/ui2/select';
+import { Switch } from '../../../components/ui2/switch';
+import BackButton from '../../../components/BackButton';
 import { Save, Loader2, Tag } from 'lucide-react';
 
 const categoryTypes: { value: CategoryType; label: string }[] = [
