+++ src/pages/admin/categories/CategoryProfile.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import { useCategoryRepository } from '../../hooks/useCategoryRepository';
-import { Card, CardHeader, CardContent } from '../../components/ui2/card';
-import { Button } from '../../components/ui2/button';
-import BackButton from '../../components/BackButton';
-import { Badge } from '../../components/ui2/badge';
+import { useCategoryRepository } from '../../../hooks/useCategoryRepository';
+import { Card, CardHeader, CardContent } from '../../../components/ui2/card';
+import { Button } from '../../../components/ui2/button';
+import BackButton from '../../../components/BackButton';
+import { Badge } from '../../../components/ui2/badge';
 import { Loader2, Tag, Pencil, Trash2 } from 'lucide-react';
 
 function CategoryProfile() {
